# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://www.baidu.com"
        },
        "version": "2.0"
    },
    "host": "192.168.1.163:7001",
    "paths": {
...
        "/v1/app": {
            "get": {
                "description": "获取app版本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "获取app版本信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ios|android",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/app/down/{type}/{name}": {
            "get": {
                "description": "app下载",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app下载链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ios|android",
                        "name": "type",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "apk name",
                        "name": "name",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/app/uploads": {
            "post": {
                "description": "更新app json版本信息,apk暂时存放服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "更新app apk暂时存放服务器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ios|android",
                        "name": "type",
                        "in": "header"
                    },
                    {
                        "type": "file",
                        "description": "file 上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool/create": {
            "post": {
                "description": "创建矿池",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "创建矿池",
                "parameters": [
                    {
                        "description": "创建矿池 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.CreateFilPool"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool/detail": {
            "get": {
                "description": "获取矿池数据详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "获取矿池数据详情",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool/list": {
            "get": {
                "description": "获取矿池数据列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "获取矿池数据列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id 可选参数",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取当页条数",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool/update": {
            "post": {
                "description": "更新矿池",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "更新矿池",
                "parameters": [
                    {
                        "description": "更新矿池 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.UpadteFilPool"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool_daily/create": {
            "post": {
                "description": "创建每日矿池收益",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "创建每日矿池收益",
                "parameters": [
                    {
                        "description": "创建每日矿池收益 需要的参数",
                        "name": "param",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.CreateFilPoolDailyIncome"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool_daily/detail": {
            "get": {
                "description": "获取每日矿池收益数据详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "获取每日矿池收益数据详情",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/v1/fil_pool_daily/list": {
            "get": {
                "description": "获取每日矿池收益数据列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "矿池"
                ],
                "summary": "获取每日矿池收益数据列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id 可选参数",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取当页条数",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateFilPool": {
            "type": "object",
            "required": [
                "effective_computing_power",
                "miner",
                "miner_available_balance",
                "miner_balance",
                "name",
                "node_id",
                "original_computing_power",
                "sector_size"
            ],
            "properties": {
                "effective_computing_power": {
                    "description": "\"有效算力\"",
                    "type": "string"
                },
                "miner": {
                    "description": "\"矿工\"",
                    "type": "string"
                },
                "miner_available_balance": {
                    "description": "\"可用余额\"",
                    "type": "number"
                },
                "miner_balance": {
                    "description": "\"账户余额\"",
                    "type": "number"
                },
                "name": {
                    "description": "\"矿池名称\"",
                    "type": "string"
                },
                "node_id": {
                    "description": "\"节点id\"",
                    "type": "string"
                },
                "original_computing_power": {
                    "description": "\"原值算力\"",
                    "type": "string"
                },
                "sector_size": {
                    "description": "\"扇区大小\"",
                    "type": "number"
                }
            }
        },
        "request.CreateFilPoolDailyIncome": {
            "type": "object",
            "required": [
                "assigned_val",
                "fil_pool_id",
                "freed",
                "last_time_val",
                "today_val"
            ],
            "properties": {
                "assigned_val": {
                    "description": "分配值",
                    "type": "number"
                },
                "fil_pool_id": {
                    "description": "矿池id",
                    "type": "string"
                },
                "freed": {
                    "description": "有效算力",
                    "type": "number"
                },
                "last_time_val": {
                    "description": "上次值",
                    "type": "number"
                },
                "today_val": {
                    "description": "本日值",
                    "type": "number"
                }
            }
        },
...
        "request.UpadteFilPool": {
            "type": "object",
            "required": [
                "effective_computing_power",
                "id",
                "miner",
                "miner_available_balance",
                "miner_balance",
                "name",
                "node_id",
                "original_computing_power",
                "sector_size"
            ],
            "properties": {
                "effective_computing_power": {
                    "description": "\"有效算力\"",
                    "type": "string"
                },
                "id": {
                    "description": "\"矿池名称\"",
                    "type": "string"
                },
                "miner": {
                    "description": "\"矿工\"",
                    "type": "string"
                },
                "miner_available_balance": {
                    "description": "\"可用余额\"",
                    "type": "number"
                },
                "miner_balance": {
                    "description": "\"账户余额\"",
                    "type": "number"
                },
                "name": {
                    "description": "\"矿池名称\"",
                    "type": "string"
                },
                "node_id": {
                    "description": "\"节点id\"",
                    "type": "string"
                },
                "original_computing_power": {
                    "description": "\"原值算力\"",
                    "type": "string"
                },
                "sector_size": {
                    "description": "\"扇区大小\"",
                    "type": "number"
                }
            }
        },
...
        "response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}